new feature: move buffer to a window
Move the current buffer to a numbered window. Use the numbered functions or the
prefix argument with the -by-number function. If prefix argument is negative (or
just present in the case of the numbered functions), the selected window becomes
the buffer's new window.

Original idea and code by @Quicknir, imported and adapted from Spacemacs.