new feature: swap buffers with a window
Swap the current buffer with a numbered window. Use the numbered functions or
the prefix argument with the -by-number function. If prefix argument is
negative (or just present in the case of the numbered functions), the input
focus follows the current buffer to the new window.

Original idea and code by @Quicknir, imported and adapted from Spacemacs.